Transparent fluid, substance color mixing now supports alpha, fluid doesn't spread upwards with empty substances, fluid side texture no longer black, improved volume calculation

+++ src/main/java/com/site21/bittermelon/common/systems/substance/Substance.java
@@ -61,8 +61,8 @@ public class Substance {
      * Calculates the molar volume of the substance in cm³/mol.
      * @return Molar volume in cm³/mol.
      */
-    public int getMolarVolume() {
-        return (int) (properties.molarMass / properties.density);
+    public float getMolarVolume() {
+        return properties.molarMass / properties.density;
     }
 
     /**
